requirements.txt
langchain-google-genai
pandas
pypdf2
python-dotenv
langgraph
pydantic
argparse